+++ src/components/MobilePairing/index.tsx
@@ -1,0 +1,11 @@
+import { lazy, ReactElement } from 'react'
+import { isPairingSupported } from 'src/logic/wallets/pairing/utils'
+import { wrapInSuspense } from 'src/utils/wrapInSuspense'
+
+const PairingDetails = lazy(() => import('src/components/AppLayout/Header/components/ProviderDetails/PairingDetails'))
+
+const MobilePairing = (): ReactElement | null => {
+  return isPairingSupported() ? wrapInSuspense(<PairingDetails />) : null
+}
+
+export default MobilePairing
